refactor: User and Role classes into the new Fan.Membership namespace #220
