+++ src/Fan.Blog/Models/Category.cs
@@ -1,6 +1,5 @@
 ﻿using Fan.Blog.Helpers;
 using Fan.Data;
-using Fan.Models;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
