+++ src/Fan.Blog/Models/Tag.cs
@@ -1,6 +1,5 @@
 ﻿using Fan.Blog.Helpers;
 using Fan.Data;
-using Fan.Models;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
