+++ src/Fan.Web/Pages/Admin/_ViewImports.cshtml
@@ -1,4 +1,4 @@
-﻿@using Fan.Models
+﻿@using Fan.Membership
 @using Fan.Settings
 @using Fan.Web.Pages.Admin
 @using Microsoft.AspNetCore.Identity
