+++ src/Fan.Web/Themes/Clarity/Views/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Identity
-@using Fan.Models
+@using Fan.Membership
 
 @inject SignInManager<User> SignInManager
 @inject UserManager<User> UserManager
