+++ src/Fan.Web/Themes/Clarity/Views/_ViewImports.cshtml
@@ -2,7 +2,7 @@
 @using Fan.Blog.Models
 @using Fan.Blog.ViewModels
 @using Fan.Blog.Services
-@using Fan.Models
+@using Fan.Membership
 @using Fan.Helpers
 @using Fan.Settings
 @using Fan.Web
