+++ src/Fan/Medias/Media.cs
@@ -1,5 +1,5 @@
 ﻿using Fan.Data;
-using Fan.Models;
+using Fan.Membership;
 using System;
 using System.ComponentModel.DataAnnotations;
