+++ test/Fan.Blog.IntegrationTests/Helpers/FakeUserManager.cs
@@ -1,4 +1,4 @@
-﻿using Fan.Models;
+﻿using Fan.Membership;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
