+++ test/Fan.Blog.UnitTests/Helpers/Actor.cs
@@ -1,4 +1,4 @@
-﻿using Fan.Models;
+﻿using Fan.Membership;
 
 namespace Fan.Blog.UnitTests.Helpers
 {
